+++ e2e/client/node/scripts/e2e.test.ts
@@ -8,13 +8,12 @@ import { startup, cleanup, shutdown } from "./e2e.start.ts";
 import type { Client } from "pg";
 import { accountTests } from "../e2e-tests/e2e.account.test.ts";
 import { generalTest } from "../e2e-tests/e2e.general.test.ts";
-import { joinAndIncrementTest, sendMintToBatcherTest } from "../e2e-tests/e2e.midnight.test.ts";
+import { joinAndIncrementTest, sendMintToBatcherTest, testMqttSubscription } from "../e2e-tests/e2e.midnight.test.ts";
 import { submitDataWithMessageAvailTest } from "../e2e-tests/e2e.avail.test.ts";
 import { testMigrations } from "../e2e-tests/e2e.migrations.ts";
 import { RPCTest } from "../e2e-tests/e2e.rpc.test.ts";
 import { tokenTests } from "../e2e-tests/e2e.tokens.ts";
 import { bitcoinTest, bitcoinBatcherTest } from "../e2e-tests/e2e.bitcoin.test.ts";
-import { testMqttSubscription } from "../e2e-tests/e2e.mqtt-subscription.test.ts";
 
 const isEnvTrue = (key: string) => ["true", "1", "yes", "y"].includes((Deno.env.get(key) || "").toLowerCase());
 
@@ -58,7 +57,7 @@ async function test() {
     );
     await joinAndIncrementTest(db, sharedState);
     await sendMintToBatcherTest(db, sharedState);
-    await testMqttSubscription(sharedState);
+    await testMqttSubscription(db, sharedState);
     await submitDataWithMessageAvailTest(db, sharedState);
     await tokenTests(db, sharedState);
     if (bitcoin_enabled) {
